Refs #112 Add primefaces and heavy refacoring
Primefaces support gives us a cheap (haha) way to add filtering
mechanism and lots of other nice GUI stuff in JavaEE.

I had to refactor most of the controller so that we aren't relying on
the data model any more as primefaces filtering seems to have issues. At
any rate this is a nicer way.

This also removes crappy pagination and uses primefaces instead.